refactor organization members sheet
